Switch include in tesseract_collision
