+++ tesseract_collision/fcl/src/fcl_utils.cpp
@@ -53,6 +53,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_PUSH
 TESSERACT_COMMON_IGNORE_WARNINGS_POP
 
 #include <tesseract_collision/fcl/fcl_utils.h>
+#include <tesseract_geometry/geometries.h>
 
 namespace tesseract_collision::tesseract_collision_fcl
 {
