+++ tesseract_geometry/include/tesseract_geometry/utils.h
@@ -26,12 +26,7 @@
 #ifndef TESSERACT_GEOMETRY_UTILS_H
 #define TESSERACT_GEOMETRY_UTILS_H
 
-#include <tesseract_common/macros.h>
-TESSERACT_COMMON_IGNORE_WARNINGS_PUSH
-#include <console_bridge/console.h>
-TESSERACT_COMMON_IGNORE_WARNINGS_POP
-
-#include <tesseract_geometry/geometries.h>
+#include <tesseract_geometry/geometry.h>
 
 namespace tesseract_geometry
 {
@@ -41,150 +36,7 @@ namespace tesseract_geometry
  * @param geom2 Second Geometry
  * @return True if identical, otherwise false
  */
-inline bool isIdentical(const Geometry& geom1, const Geometry& geom2)
-{
-  if (geom1.getType() != geom2.getType())
-    return false;
+bool isIdentical(const Geometry& geom1, const Geometry& geom2);
 
-  switch (geom1.getType())
-  {
-    case GeometryType::BOX:
-    {
-      const Box& s1 = static_cast<const Box&>(geom1);
-      const Box& s2 = static_cast<const Box&>(geom2);
-
-      if (std::abs(s1.getX() - s2.getX()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      if (std::abs(s1.getY() - s2.getY()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      if (std::abs(s1.getZ() - s2.getZ()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      break;
-    }
-    case GeometryType::SPHERE:
-    {
-      const Sphere& s1 = static_cast<const Sphere&>(geom1);
-      const Sphere& s2 = static_cast<const Sphere&>(geom2);
-
-      if (std::abs(s1.getRadius() - s2.getRadius()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      break;
-    }
-    case GeometryType::CYLINDER:
-    {
-      const Cylinder& s1 = static_cast<const Cylinder&>(geom1);
-      const Cylinder& s2 = static_cast<const Cylinder&>(geom2);
-
-      if (std::abs(s1.getRadius() - s2.getRadius()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      if (std::abs(s1.getLength() - s2.getLength()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      break;
-    }
-    case GeometryType::CONE:
-    {
-      const Cone& s1 = static_cast<const Cone&>(geom1);
-      const Cone& s2 = static_cast<const Cone&>(geom2);
-
-      if (std::abs(s1.getRadius() - s2.getRadius()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      if (std::abs(s1.getLength() - s2.getLength()) > std::numeric_limits<double>::epsilon())
-        return false;
-
-      break;
-    }
-    case GeometryType::MESH:
-    {
-      const Mesh& s1 = static_cast<const Mesh&>(geom1);
-      const Mesh& s2 = static_cast<const Mesh&>(geom2);
-
-      if (s1.getVertexCount() != s2.getVertexCount())
-        return false;
-
-      if (s1.getFaceCount() != s2.getFaceCount())
-        return false;
-
-      break;
-    }
-    case GeometryType::CONVEX_MESH:
-    {
-      const ConvexMesh& s1 = static_cast<const ConvexMesh&>(geom1);
-      const ConvexMesh& s2 = static_cast<const ConvexMesh&>(geom2);
-
-      if (s1.getVertexCount() != s2.getVertexCount())
-        return false;
-
-      if (s1.getFaceCount() != s2.getFaceCount())
-        return false;
-
-      break;
-    }
-    case GeometryType::SDF_MESH:
-    {
-      const SDFMesh& s1 = static_cast<const SDFMesh&>(geom1);
-      const SDFMesh& s2 = static_cast<const SDFMesh&>(geom2);
-
-      if (s1.getVertexCount() != s2.getVertexCount())
-        return false;
-
-      if (s1.getFaceCount() != s2.getFaceCount())
-        return false;
-
-      break;
-    }
-    case GeometryType::OCTREE:
-    {
-      const Octree& s1 = static_cast<const Octree&>(geom1);
-      const Octree& s2 = static_cast<const Octree&>(geom2);
-
-      const std::shared_ptr<const octomap::OcTree>& octree1 = s1.getOctree();
-      const std::shared_ptr<const octomap::OcTree>& octree2 = s2.getOctree();
-
-      if (octree1->getTreeType() != octree2->getTreeType())
-        return false;
-
-      if (octree1->size() != octree2->size())
-        return false;
-
-      if (octree1->getTreeDepth() != octree2->getTreeDepth())
-        return false;
-
-      if (octree1->memoryUsage() != octree2->memoryUsage())
-        return false;
-
-      if (octree1->memoryFullGrid() != octree2->memoryFullGrid())
-        return false;
-
-      break;
-    }
-    case GeometryType::POLYGON_MESH:
-    {
-      const PolygonMesh& s1 = static_cast<const PolygonMesh&>(geom1);
-      const PolygonMesh& s2 = static_cast<const PolygonMesh&>(geom2);
-
-      if (s1.getVertexCount() != s2.getVertexCount())
-        return false;
-
-      if (s1.getFaceCount() != s2.getFaceCount())
-        return false;
-
-      break;
-    }
-    default:
-    {
-      CONSOLE_BRIDGE_logError("This geometric shape type (%d) is not supported", static_cast<int>(geom1.getType()));
-      return false;
-    }
-  }
-
-  return true;
-}
 }  // namespace tesseract_geometry
 #endif
